requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aioimgur==0.1.1
aioresponses==0.7.4
atoma==0.0.17
cachetools==5.2.0
discord.py==2.1.0
diskcache==5.4.0
enkanetwork.py==1.3.3
GitPython==3.1.30
hutao_login_gateway==0.1.0
matplotlib==3.6.2
numpy==1.24.1
Pillow==9.4.0
pydantic==1.10.4
pyppeteer==1.0.2
pytest==7.2.0
python-dotenv==0.21.0
python_dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
sentry_sdk==1.12.1
six==1.16.0
uvloop==0.17.0